+++ lib/compiler.ex
@@ -6,9 +6,9 @@ defmodule Tamnoon.Compiler do
   @doc """
   Compiles the 'root.ex' component into the final file that is sent to the client.
   """
-  @spec build_from_root(ws_address :: String.t(), root :: String.t() | module()) :: :ok | {:error, atom()}
-  def build_from_root(ws_address \\ "ws://localhost:4000/ws/", root \\ Tamnoon.Components.Root) do
-    compiled_html = render_component(root, %{ws_address: ws_address})
+  @spec build_from_root(root :: String.t() | module()) :: :ok | {:error, atom()}
+  def build_from_root(root \\ Tamnoon.Components.Root) do
+    compiled_html = render_component(root)
     |> parse_tmnn_heex()
     File.write("tamnoon_out/app.html", compiled_html)
   end
@@ -33,12 +33,12 @@ defmodule Tamnoon.Compiler do
 
   def render_component(component, assigns, false) when is_binary(component) do
     "lib/components/" <> component
-    |> EEx.eval_file([r: &render_component_dyn/1, assigns: assigns])
+    |> EEx.eval_file([r: &render_component_dyn/1, h: &escape_html/1, assigns: assigns])
   end
 
   def render_component(component, assigns, false) do
     component.heex()
-    |> EEx.eval_string([r: &render_component_dyn/1, assigns: assigns])
+    |> EEx.eval_string([r: &render_component_dyn/1, h: &escape_html/1, assigns: assigns])
   end
 
   @doc """
@@ -59,19 +59,19 @@ defmodule Tamnoon.Compiler do
   """
   @spec parse_tmnn_heex(String.t()) :: String.t()
   def parse_tmnn_heex(compiled_heex) do
-    Regex.replace(~r/< *\w *[^>]*> *(?:@([a-z|\d]+))?/m, compiled_heex, fn x, innerHtml ->
-      attr_classes = get_component_attrs(x, innerHtml)
+    Regex.replace(~r/< *\w *[^>]*> *(?:@([-_a-z\d]+))?/m, compiled_heex, fn x, inner_value ->
+      attr_classes = get_component_attrs(x, inner_value)
       |> Enum.reduce("", fn {key, attrs}, acc ->
         acc <> Enum.reduce(attrs, "", fn attr, acc ->
           if (String.starts_with?(attr, "on")) do
-            acc <> " tmnnevent-" <> "#{attr}-" <> Atom.to_string(key)
+            acc <> " tmnnevent-" <> "#{attr}-" <> key
           else
-            acc <> " tmnn-" <> Atom.to_string(key) <> "-#{attr}"
+            acc <> " tmnn-" <> key <> "-#{attr}"
           end
         end)
       end)
-      x = Regex.replace(~r/(?<attr>[a-z\d]+)=@(?<key>[_a-z\d]+)/m, x, "")
-      x = Regex.replace(~r/>( *@[_a-z\d]+)/m, x, ">")
+      x = Regex.replace(~r/(?<attr>[a-z\d]+)=@(?<key>[-_a-z\d]+)/m, x, "")
+      x = Regex.replace(~r/>( *@[-_a-z\d]+)/m, x, ">")
       x = if (x =~ ~r/class="[^"]*"/ || attr_classes == "") do
         x
       else
@@ -86,14 +86,35 @@ defmodule Tamnoon.Compiler do
   end
 
   defp get_component_attrs(component, "") do
-    Regex.scan(~r/(?<attr>[a-z\d]+)=@(?<key>[_a-z\d]+)/m, component)
-    |> Enum.group_by(&(String.to_atom(Enum.at(&1, 2))), &(Enum.at(&1, 1)))
+    Regex.scan(~r/(?<attr>[a-z\d]+)=@(?<key>[-_a-z\d]+)/m, component)
+    |> Enum.group_by(&(Enum.at(&1, 2)), &(Enum.at(&1, 1)))
   end
 
-  defp get_component_attrs(component, innerHtml) do
-    [[nil, "innerHtml", innerHtml]]
+  defp get_component_attrs(component, inner_value) do
+    if (String.starts_with?(inner_value, "raw-")) do
+      [[nil, "innerHtml", String.slice(inner_value, 4..-1//1)]]
+    else
+      [[nil, "innerText", inner_value]]
+    end
     |> Enum.concat(Regex.scan(~r/(?<attr>[a-z\d]+)=@(?<key>[_a-z\d]+)/m, component))
-    |> Enum.group_by(&(String.to_atom(Enum.at(&1, 2))), &(Enum.at(&1, 1)))
+    |> Enum.group_by(&(Enum.at(&1, 2)), &(Enum.at(&1, 1)))
+  end
+
+  @escapes %{
+    "<" => "&lt;",
+    ">" => "&gt;",
+    "&" => "&amp;",
+    "\"" => "&quot;",
+    "'" => "&#39;"
+  }
+
+  @doc """
+  Escapes the content so it is not rendered as HTML. Available as `<%= h.(content) %>`
+  inside components.
+  """
+  @spec escape_html(String.t()) :: String.t()
+  def escape_html(content) do
+    String.replace(content, ["<", ">", "&", "\"", "'"], &(Map.get(@escapes, &1)))
   end
 
 end