+++ lib/root.ex
@@ -10,15 +10,13 @@ defmodule Tamnoon.Components.Root do
     <head>
       <meta name="description" content="Webpage description goes here" />
       <meta charset="utf-8">
-      <title>title</title>
+      <title>Tamnoon App</title>
       <meta name="viewport" content="width=device-width, initial-scale=1">
-      <meta name="tmnn_wsaddress" content="<%= @ws_address %>"/>
       <script src="ws_connect.js"></script>
     </head>
 
     <body>
-
-        <div class="container">
+        <div class="app-container">
             <%= r.("app.html.heex") %>
         </div>
     </body>